// packages/@fluentui-vue/next/src/components/Calendar/CalendarDay/CalendarDay.types.ts
export interface ICalendarDayStyleProps extends ICalendarDayGridStyleProps {
  /**
   * Theme provided by High-Order Component.
   */
  theme: ITheme

  /**
   * Accept custom classNames
   */
  className?: string

  // Insert CalendarDay style props below

  /**
   * Whether the header is clickable
   */
  headerIsClickable?: boolean

  /**
   * Whether week numbers are being shown
   */
  showWeekNumbers?: boolean
}

/**
 * {@docCategory Calendar}
 */
export interface ICalendarDayStyles extends Partial<ICalendarDayGridStyles> {
  /**
   * Style for the root element.
   */
  root: IStyle

  /**
   * The style for the header button and forward/back navigation button container
   */
  header: IStyle

  /**
   * The style for the title text inside the header
   */
  monthAndYear: IStyle

  /**
   * The style for the wrapper around forward/back/close buttons
   */
  monthComponents: IStyle

  /**
   * The style for the forward/back/close buttons
   */
  headerIconButton: IStyle

  /**
   * The style to apply for disabled elements
   */
  disabledStyle: IStyle
}